build: enforce type checking and linting, fix all violations
