+++ package.json
@@ -22,6 +22,7 @@
     "@mdx-js/react": "^3.1.0",
     "@next/mdx": "^15.0.2",
     "@radix-ui/react-avatar": "^1.0.4",
+    "@radix-ui/react-dialog": "^1.1.18",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-separator": "^1.1.0",
     "@radix-ui/react-slot": "^1.0.2",
@@ -32,6 +33,9 @@
     "framer-motion": "^11.2.10",
     "gray-matter": "^4.0.3",
     "lucide-react": "^0.395.0",
+    "mdast-util-from-markdown": "^2.0.3",
+    "mdast-util-gfm": "^3.1.0",
+    "micromark-extension-gfm": "^3.0.0",
     "next": "14.2.4",
     "next-mdx-remote": "^5.0.0",
     "next-themes": "^0.3.0",
@@ -57,6 +61,7 @@
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "@types/sharp": "^0.32.0",
+    "chokidar": "^5.0.0",
     "commander": "^12.1.0",
     "concurrently": "^9.1.0",
     "cross-env": "^7.0.3",
