+++ src/app/blog/tag/[tag]/page.tsx
@@ -1,11 +1,11 @@
 import { getBlogPosts } from "@/data/blog";
-import BlogPage from "../../page";
+import { BlogList } from "@/components/blog/BlogList";
 
 // Add this function to generate static paths for all tags
 export async function generateStaticParams() {
   const posts = await getBlogPosts();
   const tags = new Set(posts.flatMap(post => post.metadata.tags ?? []));
-  
+
   return Array.from(tags).map((tag) => ({
     tag: tag,
   }));
@@ -13,9 +13,9 @@ export async function generateStaticParams() {
 
 export default async function TagPage({ params }: { params: { tag: string } }) {
   const posts = await getBlogPosts();
-  const filteredPosts = posts.filter(post => 
+  const filteredPosts = posts.filter(post =>
     post.metadata.tags?.includes(params.tag)
   );
-  
-  return <BlogPage posts={filteredPosts} tag={params.tag} />;
-} 
+
+  return <BlogList posts={filteredPosts} tag={params.tag} />;
+}
