+++ src/pages/_app.tsx
@@ -1,6 +1,7 @@
 import { CustomMDXProvider } from '@/components/mdx/MDXProvider'
+import type { AppProps } from 'next/app'
 
-export default function App({ Component, pageProps }) {
+export default function App({ Component, pageProps }: AppProps) {
   return (
     <CustomMDXProvider>
       <Component {...pageProps} />
